requirements.txt
ipython
requests
flask==1.1.2
python-dotenv==0.18.0